requirements.txt
PyYAML
jsonschema
requests
beautifulsoup4
rembg
onnxruntime
opencv-python
streamlit
